+++ flink-state-backends/flink-statebackend-rocksdb/src/main/java/org/apache/flink/contrib/streaming/state/RocksDBNativeMetricOptions.java
@@ -18,6 +18,8 @@
 
 package org.apache.flink.contrib.streaming.state;
 
+import org.apache.flink.configuration.ConfigOption;
+import org.apache.flink.configuration.ConfigOptions;
 import org.apache.flink.configuration.Configuration;
 
 import java.io.Serializable;
@@ -36,100 +38,211 @@ import java.util.Set;
  */
 public class RocksDBNativeMetricOptions implements Serializable {
 
-	private Set<String> properties;
+	public static final ConfigOption<Boolean> MONITOR_NUM_IMMUTABLE_MEM_TABLES = ConfigOptions
+		.key(RocksDBProperty.NumImmutableMemTable.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the number of immutable memtables in RocksDB.");
 
+	public static final ConfigOption<Boolean> MONITOR_MEM_TABLE_FLUSH_PENDING = ConfigOptions
+		.key(RocksDBProperty.MemTableFlushPending.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the number of pending memtable flushes in RocksDB.");
+
+	public static final ConfigOption<Boolean> TRACK_COMPACTION_PENDING = ConfigOptions
+		.key(RocksDBProperty.CompactionPending.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Track pending compactions in RocksDB. Returns 1 if a compaction is pending, 0 otherwise.");
+
+	public static final ConfigOption<Boolean> MONITOR_BACKGROUND_ERRORS = ConfigOptions
+		.key(RocksDBProperty.BackgroundErrors.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the number of background errors in RocksDB.");
+
+	public static final ConfigOption<Boolean> MONITOR_CUR_SIZE_ACTIVE_MEM_TABLE = ConfigOptions
+		.key(RocksDBProperty.CurSizeActiveMemTable.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the approximate size of the active memtable in bytes.");
+
+	public static final ConfigOption<Boolean> MONITOR_CUR_SIZE_ALL_MEM_TABLE = ConfigOptions
+		.key(RocksDBProperty.CurSizeAllMemTables.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the approximate size of the active and unflushed immutable memtables" +
+			" in bytes.");
+
+	public static final ConfigOption<Boolean> MONITOR_SIZE_ALL_MEM_TABLES = ConfigOptions
+		.key(RocksDBProperty.SizeAllMemTables.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the approximate size of the active, unflushed immutable, " +
+			"and pinned immutable memtables in bytes.");
+
+	public static final ConfigOption<Boolean> MONITOR_NUM_ENTRIES_ACTIVE_MEM_TABLE = ConfigOptions
+		.key(RocksDBProperty.NumEntriesActiveMemTable.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the total number of entries in the active memtable.");
+
+	public static final ConfigOption<Boolean> MONITOR_NUM_ENTRIES_IMM_MEM_TABLES = ConfigOptions
+		.key(RocksDBProperty.NumEntriesImmMemTables.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the total number of entries in the unflushed immutable memtables.");
+
+	public static final ConfigOption<Boolean> MONITOR_NUM_DELETES_ACTIVE_MEM_TABLE = ConfigOptions
+		.key(RocksDBProperty.NumDeletesActiveMemTable.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the total number of delete entries in the active memtable.");
+
+	public static final ConfigOption<Boolean> MONITOR_NUM_DELETES_IMM_MEM_TABLE = ConfigOptions
+		.key(RocksDBProperty.NumDeletesImmMemTables.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the total number of delete entries in the unflushed immutable memtables.");
+
+	public static final ConfigOption<Boolean> ESTIMATE_NUM_KEYS = ConfigOptions
+		.key(RocksDBProperty.EstimateNumKeys.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Estimate the number of keys in RocksDB.");
+
+	public static final ConfigOption<Boolean> ESTIMATE_TABLE_READERS_MEM = ConfigOptions
+		.key(RocksDBProperty.EstimateTableReadersMem.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Estimate the memory used for reading SST tables, excluding memory" +
+			" used in block cache (e.g.,filter and index blocks) in bytes.");
+
+	public static final ConfigOption<Boolean> MONITOR_NUM_SNAPSHOTS = ConfigOptions
+		.key(RocksDBProperty.NumSnapshots.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the number of unreleased snapshots of the database.");
+
+	public static final ConfigOption<Boolean> MONITOR_NUM_LIVE_VERSIONS = ConfigOptions
+		.key(RocksDBProperty.NumLiveVersions.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor number of live versions. Version is an internal data structure. " +
+			"See RocksDB file version_set.h for details. More live versions often mean more SST files are held " +
+			"from being deleted, by iterators or unfinished compactions.");
+
+	public static final ConfigOption<Boolean> ESTIMATE_LIVE_DATA_SIZE = ConfigOptions
+		.key(RocksDBProperty.EstimateLiveDataSize.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Estimate of the amount of live data in bytes.");
+
+	public static final ConfigOption<Boolean> MONITOR_TOTAL_SST_FILES_SIZE = ConfigOptions
+		.key(RocksDBProperty.TotalSstFilesSize.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the total size (bytes) of all SST files." +
+			"WARNING: may slow down online queries if there are too many files.");
+
+	public static final ConfigOption<Boolean> ESTIMATE_PENDING_COMPACTION_BYTES = ConfigOptions
+		.key(RocksDBProperty.EstimatePendingCompactionBytes.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Estimated total number of bytes compaction needs to rewrite to get all levels " +
+			"down to under target size. Not valid for other compactions than level-based.");
+
+	public static final ConfigOption<Boolean> MONITOR_NUM_RUNNING_COMPACTIONS = ConfigOptions
+		.key(RocksDBProperty.NumRunningCompactions.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the number of currently running compactions.");
+
+	public static final ConfigOption<Boolean> MONITOR_NUM_RUNNING_FLUSHES = ConfigOptions
+		.key(RocksDBProperty.NumRunningFlushes.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the number of currently running flushes.");
+
+	public static final ConfigOption<Boolean> MONITOR_ACTUAL_DELAYED_WRITE_RATE = ConfigOptions
+		.key(RocksDBProperty.ActualDelayedWriteRate.getConfigKey())
+		.defaultValue(false)
+		.withDescription("Monitor the current actual delayed write rate. 0 means no delay.");
 	/**
 	 * Creates a {@link RocksDBNativeMetricOptions} based on an
 	 * external configuration.
 	 */
 	public static RocksDBNativeMetricOptions fromConfig(Configuration config) {
 		RocksDBNativeMetricOptions options = new RocksDBNativeMetricOptions();
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_IMMUTABLE_MEM_TABLES)) {
+		if (config.getBoolean(MONITOR_NUM_IMMUTABLE_MEM_TABLES)) {
 			options.enableNumImmutableMemTable();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_MEM_TABLE_FLUSH_PENDING)) {
+		if (config.getBoolean(MONITOR_MEM_TABLE_FLUSH_PENDING)) {
 			options.enableMemTableFlushPending();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_COMPACTION_PENDING)) {
+		if (config.getBoolean(TRACK_COMPACTION_PENDING)) {
 			options.enableCompactionPending();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_BACKGROUND_ERRORS)) {
+		if (config.getBoolean(MONITOR_BACKGROUND_ERRORS)) {
 			options.enableBackgroundErrors();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_CUR_SIZE_ACTIVE_MEM_TABLE)) {
+		if (config.getBoolean(MONITOR_CUR_SIZE_ACTIVE_MEM_TABLE)) {
 			options.enableCurSizeActiveMemTable();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_CUR_SIZE_ALL_MEM_TABLE)) {
+		if (config.getBoolean(MONITOR_CUR_SIZE_ALL_MEM_TABLE)) {
 			options.enableCurSizeAllMemTables();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_SIZE_ALL_MEM_TABLES)) {
+		if (config.getBoolean(MONITOR_SIZE_ALL_MEM_TABLES)) {
 			options.enableSizeAllMemTables();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_ENTRIES_ACTIVE_MEM_TABLE)) {
+		if (config.getBoolean(MONITOR_NUM_ENTRIES_ACTIVE_MEM_TABLE)) {
 			options.enableNumEntriesActiveMemTable();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_ENTRIES_IMM_MEM_TABLES)) {
+		if (config.getBoolean(MONITOR_NUM_ENTRIES_IMM_MEM_TABLES)) {
 			options.enableNumEntriesImmMemTables();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_DELETES_ACTIVE_MEM_TABLE)) {
+		if (config.getBoolean(MONITOR_NUM_DELETES_ACTIVE_MEM_TABLE)) {
 			options.enableNumDeletesActiveMemTable();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_DELETES_IMM_MEM_TABLE)) {
+		if (config.getBoolean(MONITOR_NUM_DELETES_IMM_MEM_TABLE)) {
 			options.enableNumDeletesImmMemTables();
 		}
 
-		if (config.getBoolean(RocksDBOptions.ESTIMATE_NUM_KEYS)) {
+		if (config.getBoolean(ESTIMATE_NUM_KEYS)) {
 			options.enableEstimateNumKeys();
 		}
 
-		if (config.getBoolean(RocksDBOptions.ESTIMATE_TABLE_READERS_MEM)) {
+		if (config.getBoolean(ESTIMATE_TABLE_READERS_MEM)) {
 			options.enableEstimateTableReadersMem();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_SNAPSHOTS)) {
+		if (config.getBoolean(MONITOR_NUM_SNAPSHOTS)) {
 			options.enableNumSnapshots();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_LIVE_VERSIONS)) {
+		if (config.getBoolean(MONITOR_NUM_LIVE_VERSIONS)) {
 			options.enableNumLiveVersions();
 		}
 
-		if (config.getBoolean(RocksDBOptions.ESTIMATE_LIVE_DATA_SIZE)) {
+		if (config.getBoolean(ESTIMATE_LIVE_DATA_SIZE)) {
 			options.enableEstimateLiveDataSize();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_TOTAL_SST_FILES_SIZE)) {
+		if (config.getBoolean(MONITOR_TOTAL_SST_FILES_SIZE)) {
 			options.enableTotalSstFilesSize();
 		}
 
-		if (config.getBoolean(RocksDBOptions.ESTIMATE_PENDING_COMPACTION_BYTES)) {
+		if (config.getBoolean(ESTIMATE_PENDING_COMPACTION_BYTES)) {
 			options.enableEstimatePendingCompactionBytes();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_RUNNING_COMPACTIONS)) {
+		if (config.getBoolean(MONITOR_NUM_RUNNING_COMPACTIONS)) {
 			options.enableNumRunningCompactions();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_NUM_RUNNING_FLUSHES)) {
+		if (config.getBoolean(MONITOR_NUM_RUNNING_FLUSHES)) {
 			options.enableNumRunningFlushes();
 		}
 
-		if (config.getBoolean(RocksDBOptions.MONITOR_ACTUAL_DELAYED_WRITE_RATE)) {
+		if (config.getBoolean(MONITOR_ACTUAL_DELAYED_WRITE_RATE)) {
 			options.enableActualDelayedWriteRate();
 		}
 
 		return options;
 	}
+
+	private Set<String> properties;
 
 	public RocksDBNativeMetricOptions() {
 		this.properties = new HashSet<>();
