+++ flink-state-backends/flink-statebackend-rocksdb/src/test/java/org/apache/flink/contrib/streaming/state/RocksDBStateBackendConfigTest.java
@@ -502,29 +502,6 @@ public class RocksDBStateBackendConfigTest {
 	}
 
 	// ------------------------------------------------------------------------
-	//  Native Metrics
-	// ------------------------------------------------------------------------
-
-	@Test
-	public void testNativeMetricsConfigurable() {
-		for (RocksDBProperty property : RocksDBProperty.values()) {
-			Configuration config = new Configuration();
-			config.setBoolean(property.getConfigKey(), true);
-
-			RocksDBNativeMetricOptions options = RocksDBNativeMetricOptions.fromConfig(config);
-
-			Assert.assertTrue(
-				String.format("Failed to enable native metrics with property %s", property.getConfigKey()),
-				options.isEnabled());
-
-			Assert.assertTrue(
-				String.format("Failed to enable native metric %s using config", property.getConfigKey()),
-				options.getProperties().contains(property.getRocksDBProperty())
-			);
-		}
-	}
-
-	// ------------------------------------------------------------------------
 	//  Utilities
 	// ------------------------------------------------------------------------
